Fix a flaky test based on #4204
